# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SQL Practice API
  version: "1.0.0"
  description: |
    API for the SQL Practice app (interactive SQL tasks checked against a real
    PostgreSQL database, sessions/learning plans, login/role system).

    **This spec is hand-maintained, not generated from code.** It documents a
    first, deliberately partial slice of the API (see "Phase 1" scope below) —
    if anything here looks stale compared to the actual route code, trust the
    code and update this file.

    ### Authentication
    This API uses a session cookie (`connect.sid`, set by `express-session` +
    `connect-pg-simple`), **not** bearer tokens. To use "Try it out" below:
    1. Call `POST /api/auth/login` with a real username/password — a valid
       session cookie is stored by your browser automatically.
    2. Every subsequent "Try it out" call from this same page reuses that
       cookie automatically — no token to copy/paste anywhere.
    3. Call `POST /api/auth/logout` to end the session.

    This only works because Swagger UI is served by the backend itself
    (same origin as `/api/*`) — see `src/routes/docs.js`.

    ### Phase 1 scope
    Only a first, "safe to explore" slice of the API is documented here:
    health, auth, datasets, tasks (read + solution), the query sandbox, and
    the table browser. Endpoints with heavier role/ownership matrices
    (sessions, users, progress, mentor assignments, mentor students) are
...
    full API surface in the meantime.
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: >-
        Session cookie set by POST /api/auth/login. httpOnly, sameSite=lax,
        secure in production. There is no bearer-token alternative.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required: [error]
      example:
        error: "Not authenticated."
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, mentor, student]
          description: >-
            DB value is always one of these three. "mentor" is shown in the
            frontend UI as "Professor" — the API value itself is unaffected.
      required: [id, username, role]
      example:
        id: 1
        username: default
        role: student
...
    MentorAssignment:
      type: object
      properties:
        id: { type: integer }
        mentor_id: { type: integer }
        mentor_username: { type: string }
        mentor_role: { type: string, enum: [mentor] }
        student_id: { type: integer }
        student_username: { type: string }
        student_role: { type: string, enum: [student] }
        created_at: { type: string, format: date-time }
      example:
        id: 3
        mentor_id: 2
        mentor_username: prof
        mentor_role: mentor
        student_id: 7
        student_username: jane_student
        student_role: student
        created_at: "2026-05-01T12:00:00.000Z"
...
security:
  - sessionCookie: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in and start a session
      description: >-
        On success, regenerates the session (fixation mitigation) and sets
        the `connect.sid` cookie. Returns the same generic error message
        whether the username doesn't exist, has no password set yet, or the
        password is wrong — this deliberately does not reveal which case
        occurred.
      operationId: login
      security: []
      x-visibility: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string, format: password }
              required: [username, password]
            example:
              username: default
              password: change-me-please
      responses:
        "200":
          description: Login succeeded; the session cookie is now set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Missing username or password.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Invalid username or password (generic message either way).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: "Invalid username or password."
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the currently logged-in user
      description: >-
        Reads `id`/`username`/`role` fresh from the database on every call
        (nothing but `userId` is cached in the session) — a role change or
        account deletion takes effect immediately on the next call.
      operationId: getMe
      x-visibility: authenticated
      responses:
        "200":
          description: The authenticated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: No active session, or the session's user no longer exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: "Not authenticated."
...
  /api/mentor-assignments:
    get:
      tags: [Mentor Assignments]
      summary: List all mentor-student assignments
      description: Admin-only. Joined with username/role for both sides; never selects password_hash.
      operationId: listMentorAssignments
      x-required-role: admin
      x-visibility: admin-any
      responses:
        "200":
          description: All assignments, ordered by mentor then student username.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MentorAssignment" }
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Authenticated but not an admin.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    post:
      tags: [Mentor Assignments]
      summary: Assign a mentor to a student
      description: >-
        Admin-only. `mentorId` must belong to a user with role "mentor",
        `studentId` must belong to a user with role "student" (400
        otherwise) — this immediately grants the mentor canAccessStudent
        access to that student's sessions/progress (checked live on every
        request, no caching). **Re-assigning an already-assigned pair is
        treated as an idempotent success** (200 with the existing row), not
        a 409 conflict — the desired end state already exists.
      operationId: createMentorAssignment
      x-required-role: admin
      x-visibility: admin-any
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mentorId: { type: integer }
                studentId: { type: integer }
              required: [mentorId, studentId]
      responses:
        "200":
          description: Assignment already existed — returned as-is (idempotent, not an error).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MentorAssignment" }
        "201":
          description: Assignment created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MentorAssignment" }
        "400":
          description: Missing mentorId/studentId, mentorId equals studentId, or the ids don't have the matching mentor/student role.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: 'mentorId must belong to a user with role "mentor".'
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Authenticated but not an admin.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: mentorId or studentId does not refer to an existing user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/mentor-assignments/{id}:
    delete:
      tags: [Mentor Assignments]
      summary: Remove a mentor-student assignment
      description: >-
        Admin-only. Deletes only the assignment row — never touches either
        user account. **Immediately revokes the mentor's access** to that
        student's sessions/progress on the mentor's very next request, since
        canAccessStudent checks mentor_assignments live, with nothing cached.
      operationId: deleteMentorAssignment
      x-required-role: admin
      x-visibility: admin-any
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Assignment removed.
          content:
            application/json:
              schema:
                type: object
                properties: { success: { type: boolean, enum: [true] } }
        "400":
          description: Invalid assignment id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Authenticated but not an admin.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Assignment not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...